TinyMT: Header-only
Move implementation of init to header.
